parse: Handle leading ? on strings (#20)
[`URL#search`][1] returns a string with a leading `?` character.
If you pass the `search` of a URL directly to rackstring, e.g.
`parse(window.location.search)`, rackstring will currently include that
character as part of the name of the first key parsed out of the query
string. E.g. `?foo=bar` ends up as `{ "?foo": "bar" }`.

Node's built-in `querystring` does the same thing, so if compatability
with that interface is paramount then this probably shouldn't be
changed. But since I think the most common use cases of `parse` is
probably to parse a `search` string from a `URL` instance or `Location`
instance, this seems like a footgun that might be worth avoiding.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/URL/search